+++ src/nodes/block.rs
@@ -4,6 +4,7 @@ use crate::nodes::{LastStatement, ReturnStatement, Statement, Token};
 pub struct BlockTokens {
     pub semicolons: Vec<Option<Token>>,
     pub last_semicolon: Option<Token>,
+    pub final_token: Option<Token>,
 }
 
 impl BlockTokens {
@@ -14,6 +15,9 @@ impl BlockTokens {
         if let Some(last_semicolon) = &mut self.last_semicolon {
             last_semicolon.clear_comments();
         }
+        if let Some(final_token) = &mut self.final_token {
+            final_token.clear_comments();
+        }
     }
 
     pub fn clear_whitespaces(&mut self) {
@@ -22,6 +26,9 @@ impl BlockTokens {
         }
         if let Some(last_semicolon) = &mut self.last_semicolon {
             last_semicolon.clear_whitespaces();
+        }
+        if let Some(final_token) = &mut self.final_token {
+            final_token.clear_whitespaces();
         }
     }
 
@@ -32,6 +39,9 @@ impl BlockTokens {
         if let Some(last_semicolon) = &mut self.last_semicolon {
             last_semicolon.replace_referenced_tokens(code);
         }
+        if let Some(final_token) = &mut self.final_token {
+            final_token.replace_referenced_tokens(code);
+        }
     }
 
     pub(crate) fn shift_token_line(&mut self, amount: usize) {
@@ -40,6 +50,9 @@ impl BlockTokens {
         }
         if let Some(last_semicolon) = &mut self.last_semicolon {
             last_semicolon.shift_token_line(amount);
+        }
+        if let Some(final_token) = &mut self.final_token {
+            final_token.shift_token_line(amount);
         }
     }
 }
@@ -73,6 +86,11 @@ impl Block {
     #[inline]
     pub fn get_tokens(&self) -> Option<&BlockTokens> {
         self.tokens.as_ref().map(|tokens| tokens.as_ref())
+    }
+
+    #[inline]
+    pub fn mutate_tokens(&mut self) -> Option<&mut BlockTokens> {
+        self.tokens.as_mut().map(|tokens| tokens.as_mut())
     }
 
     pub fn push_statement<T: Into<Statement>>(&mut self, statement: T) {
@@ -428,7 +446,8 @@ mod test {
             block.get_tokens(),
             Some(&BlockTokens {
                 semicolons: vec![None],
-                last_semicolon: None
+                last_semicolon: None,
+                final_token: None,
             })
         );
     }
@@ -440,6 +459,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: vec![Some(Token::from_content(";"))],
                 last_semicolon: None,
+                final_token: None,
             });
         block.clear();
 
@@ -453,6 +473,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: Vec::new(),
                 last_semicolon: Some(Token::from_content(";")),
+                final_token: None,
             });
         block.clear();
 
@@ -466,6 +487,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: vec![Some(Token::from_content(";"))],
                 last_semicolon: None,
+                final_token: None,
             });
         block.set_statements(Vec::new());
 
@@ -479,6 +501,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: Vec::new(),
                 last_semicolon: Some(Token::from_content(";")),
+                final_token: None,
             });
 
         assert_eq!(
@@ -497,6 +520,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: vec![Some(Token::from_content(";"))],
                 last_semicolon: None,
+                final_token: None,
             });
 
         block.filter_statements(|_statement| false);
@@ -506,6 +530,7 @@ mod test {
             Block::default().with_tokens(BlockTokens {
                 semicolons: Vec::new(),
                 last_semicolon: None,
+                final_token: None,
             })
         );
     }
@@ -518,6 +543,7 @@ mod test {
             .with_tokens(BlockTokens {
                 semicolons: vec![Some(Token::from_content(";"))],
                 last_semicolon: None,
+                final_token: None,
             });
 
         block.filter_mut_statements(|_statement| false);
@@ -527,6 +553,7 @@ mod test {
             Block::default().with_tokens(BlockTokens {
                 semicolons: Vec::new(),
                 last_semicolon: None,
+                final_token: None,
             })
         );
     }